Start heuristics
Wił and I just talked about how to add onto heuristics from this dictionary data-structure. I assigned a value, kinda based on reverse Scrabble values, to each letter. We're not sure what to do about black squares, but for now, we're adding a zero to the cell's value that has a black square.